daemons/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
aiosqlite>=0.20
sse-starlette>=2.0
pyyaml>=6.0
httpx>=0.27
